% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{example_queries}
\alias{example_queries}
\alias{query1}
\alias{query2}
\alias{query3}
\title{Example Europe PMC queries for package vignettes.}
\format{
A character string containing a Europe PMC advanced search query.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
query1

query2

query3
}
\description{
Example Europe PMC queries for package vignettes.
}
\keyword{datasets}
