% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{get_article_data}
\alias{get_article_data}
\title{Get Article Data}
\usage{
get_article_data(
  queries,
  limit = 7500,
  min_year = NULL,
  max_year = NULL,
  min_citations = 5,
  n_cores = default_worker_count()
)
}
\arguments{
\item{queries}{Input a Europe PMC query}

\item{limit}{Minimum number of articles to retrieve}

\item{min_year}{Minimum publication year}

\item{max_year}{Maximum publication year}

\item{min_citations}{Minimum number of citations per article}

\item{n_cores}{Number of cores to use across multiple queries.}
}
\value{
A tibble of the article data
}
\description{
This function retrieves article data from europe pmc.
}
