% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{get_articles_with_journal_data}
\alias{get_articles_with_journal_data}
\title{Get Articles With Journal Data}
\usage{
get_articles_with_journal_data(article_data, journal_data)
}
\arguments{
\item{article_data}{DESCRIPTION.}

\item{journal_data}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
