% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journals.R
\name{get_journal_data}
\alias{get_journal_data}
\title{Get Journal Data}
\usage{
get_journal_data(data = "incities")
}
\arguments{
\item{data}{A data source included in the package's data folder. Can be scimago or incities.}
}
\value{
A tibble of the journal data.
}
\description{
This function retrieves journal data from an existing data file.
}
