% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{get_unique_issns}
\alias{get_unique_issns}
\title{Get Unique ISSNs}
\usage{
get_unique_issns(issns)
}
\arguments{
\item{issns}{The issns from the articles}
}
\value{
A tibble of filtered issns
}
\description{
This function filters ISSNs.
}
