% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_word_cloud}
\alias{get_word_cloud}
\title{Get Word Cloud}
\usage{
get_word_cloud(pubmed_ids, plot_name)
}
\arguments{
\item{pubmed_ids}{A list of pubmed ids}

\item{plot_name}{Path to the output word cloud}
}
\description{
This function imports a list of pubmed ids and creates a word cloud.
}
