% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{install_journalanalysis_packages}
\alias{install_journalanalysis_packages}
\title{Install JournalAnalysis Packages}
\usage{
install_journalanalysis_packages()
}
\description{
This function helps user to install packages used in this package
}
