% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{issn_to_article_data}
\alias{issn_to_article_data}
\title{ISSN to Article Data}
\usage{
issn_to_article_data(data, issns)
}
\arguments{
\item{data}{The data tibble of the articles}

\item{issns}{The issns from the articles}
}
\value{
A tibble of the articles filtered by ISSNs
}
\description{
This function filters article data by ISSN.
}
