% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journals.R
\name{issn_to_journal_data}
\alias{issn_to_journal_data}
\title{ISSN to Journal Data}
\usage{
issn_to_journal_data(data, issns)
}
\arguments{
\item{data}{The data source of the journal}

\item{issns}{The issns from the journal source}
}
\value{
A tibble of the data filtered by ISSN
}
\description{
This function retrieves journal data from an existing data file.
}
