% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save as CSV}
\usage{
save_as_csv(data, filename)
}
\arguments{
\item{data}{The dataframe or tibble}

\item{filename}{Name or path of output file without .csv extension}
}
\description{
This function saves a tibble or dataframe as a csv file.
}
