% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\name{get_publication_data}
\alias{get_publication_data}
\title{Get Publication Data}
\usage{
get_publication_data(journal_source, queries, limit = 7500,
  min_year = 2008, max_year = 2018, min_citations = 5)
}
\arguments{
\item{journal_source}{The journal to search}

\item{queries}{Use one or multiple queries}

\item{limit}{A minimum number of articles}

\item{min_year}{Minimum year to search}

\item{max_year}{Maximum year to search}

\item{min_citations}{Minimum number of citations in a journal}
}
\value{
A list of data objects including journals, articles, and combined data.
}
\description{
This function uses a journal source to combine data based on queries.
}
